Extract command execution logic
