+++ receiver/.gitignore
@@ -33,3 +33,5 @@ bin/*
 tmp
 
 vendor
+
+/receiver
